pyproject.toml
[project]
name = "speaking-exercises"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "pyyaml>=6.0.2",
]
